requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile '.\requirements.in'
#
aiofiles==0.5.0           # via -r .\requirements.in
aiohttp==3.6.2            # via -r .\requirements.in
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp
chardet==3.0.4            # via aiohttp
idna==2.9                 # via yarl
multidict==4.7.6          # via aiohttp, yarl
pillow==7.1.2             # via -r .\requirements.in
pypiwin32==223            # via win10toast
pywin32==228              # via pypiwin32
win10toast==0.9           # via -r .\requirements.in
yarl==1.4.2               # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
